Move syscall related stuff to syscall.py
